refactor and implement timestampadd
