+++ sql/expression/function/date.go
@@ -29,300 +29,6 @@ import (
 	"github.com/dolthub/go-mysql-server/sql/types"
 )
 
-// NewAddDate returns a new function expression, or an error if one couldn't be created. The ADDDATE
-// function is a synonym for DATE_ADD, with the one exception that if the second argument is NOT an
-// explicitly declared interval, then the value is used and the interval period is assumed to be DAY.
-// In either case, this function will actually return a *DateAdd struct.
-func NewAddDate(args ...sql.Expression) (sql.Expression, error) {
-	if len(args) != 2 {
-		return nil, sql.ErrInvalidArgumentNumber.New("ADDDATE", 2, len(args))
-	}
-
-	// If the interval is explicitly specified, then we simply pass it all to DateSub
-	i, ok := args[1].(*expression.Interval)
-	if ok {
-		return &DateAdd{args[0], i}, nil
-	}
-
-	// Otherwise, the interval period is assumed to be DAY
-	i = expression.NewInterval(args[1], "DAY")
-	return &DateAdd{args[0], i}, nil
-}
-
-// DateAdd adds an interval to a date.
-type DateAdd struct {
-	Date     sql.Expression
-	Interval *expression.Interval
-}
-
-var _ sql.FunctionExpression = (*DateAdd)(nil)
-var _ sql.CollationCoercible = (*DateAdd)(nil)
-
-// NewDateAdd creates a new date add function.
-func NewDateAdd(args ...sql.Expression) (sql.Expression, error) {
-	if len(args) != 2 {
-		return nil, sql.ErrInvalidArgumentNumber.New("DATE_ADD", 2, len(args))
-	}
-
-	i, ok := args[1].(*expression.Interval)
-	if !ok {
-		return nil, fmt.Errorf("DATE_ADD expects an interval as second parameter")
-	}
-
-	return &DateAdd{args[0], i}, nil
-}
-
-// FunctionName implements sql.FunctionExpression
-func (d *DateAdd) FunctionName() string {
-	return "date_add"
-}
-
-// Description implements sql.FunctionExpression
-func (d *DateAdd) Description() string {
-	return "adds the interval to the given date."
-}
-
-// Children implements the sql.Expression interface.
-func (d *DateAdd) Children() []sql.Expression {
-	return []sql.Expression{d.Date, d.Interval}
-}
-
-// Resolved implements the sql.Expression interface.
-func (d *DateAdd) Resolved() bool {
-	return d.Date.Resolved() && d.Interval.Resolved()
-}
-
-// IsNullable implements the sql.Expression interface.
-func (d *DateAdd) IsNullable() bool {
-	return true
-}
-
-// Type implements the sql.Expression interface.
-func (d *DateAdd) Type() sql.Type {
-	sqlType := dateOffsetType(d.Date, d.Interval)
-	return sqlType
-}
-
-// CollationCoercibility implements the interface sql.CollationCoercible.
-func (*DateAdd) CollationCoercibility(ctx *sql.Context) (collation sql.CollationID, coercibility byte) {
-	return ctx.GetCollation(), 4
-}
-
-// WithChildren implements the Expression interface.
-func (d *DateAdd) WithChildren(children ...sql.Expression) (sql.Expression, error) {
-	return NewDateAdd(children...)
-}
-
-// Eval implements the sql.Expression interface.
-func (d *DateAdd) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
-	date, err := d.Date.Eval(ctx, row)
-	if err != nil {
-		return nil, err
-	}
-
-	if date == nil {
-		return nil, nil
-	}
-
-	delta, err := d.Interval.EvalDelta(ctx, row)
-	if err != nil {
-		return nil, err
-	}
-
-	if delta == nil {
-		return nil, nil
-	}
-
-	var dateVal interface{}
-	dateVal, _, err = types.DatetimeMaxPrecision.Convert(ctx, date)
-	if err != nil {
-		ctx.Warn(1292, err.Error())
-		return nil, nil
-	}
-
-	// return appropriate type
-	res := types.ValidateTime(delta.Add(dateVal.(time.Time)))
-	if res == nil {
-		return nil, nil
-	}
-
-	resType := d.Type()
-	if types.IsText(resType) {
-		// If the input is a properly formatted date/datetime string, the output should also be a string
-		if dateStr, isStr := date.(string); isStr {
-			if res.(time.Time).Nanosecond() > 0 {
-				return res.(time.Time).Format(sql.DatetimeLayoutNoTrim), nil
-			}
-			if isHmsInterval(d.Interval) {
-				return res.(time.Time).Format(sql.TimestampDatetimeLayout), nil
-			}
-			for _, layout := range types.DateOnlyLayouts {
-				if _, pErr := time.Parse(layout, dateStr); pErr != nil {
-					continue
-				}
-				return res.(time.Time).Format(sql.DateLayout), nil
-			}
-		}
-	}
-
-	ret, _, err := resType.Convert(ctx, res)
-	if err != nil {
-		return nil, err
-	}
-	return ret, nil
-}
-
-func (d *DateAdd) String() string {
-	return fmt.Sprintf("%s(%s,%s)", d.FunctionName(), d.Date, d.Interval)
-}
-
-// NewSubDate returns a new function expression, or an error if one couldn't be created. The SUBDATE
-// function is a synonym for DATE_SUB, with the one exception that if the second argument is NOT an
-// explicitly declared interval, then the value is used and the interval period is assumed to be DAY.
-// In either case, this function will actually return a *DateSub struct.
-func NewSubDate(args ...sql.Expression) (sql.Expression, error) {
-	if len(args) != 2 {
-		return nil, sql.ErrInvalidArgumentNumber.New("SUBDATE", 2, len(args))
-	}
-
-	// If the interval is explicitly specified, then we simply pass it all to DateSub
-	i, ok := args[1].(*expression.Interval)
-	if ok {
-		return &DateSub{args[0], i}, nil
-	}
-
-	// Otherwise, the interval period is assumed to be DAY
-	i = expression.NewInterval(args[1], "DAY")
-	return &DateSub{args[0], i}, nil
-}
-
-// DateSub subtracts an interval from a date.
-type DateSub struct {
-	Date     sql.Expression
-	Interval *expression.Interval
-}
-
-var _ sql.FunctionExpression = (*DateSub)(nil)
-var _ sql.CollationCoercible = (*DateSub)(nil)
-
-// NewDateSub creates a new date add function.
-func NewDateSub(args ...sql.Expression) (sql.Expression, error) {
-	if len(args) != 2 {
-		return nil, sql.ErrInvalidArgumentNumber.New("DATE_SUB", 2, len(args))
-	}
-
-	i, ok := args[1].(*expression.Interval)
-	if !ok {
-		return nil, fmt.Errorf("DATE_SUB expects an interval as second parameter")
-	}
-
-	return &DateSub{args[0], i}, nil
-}
-
-// FunctionName implements sql.FunctionExpression
-func (d *DateSub) FunctionName() string {
-	return "date_sub"
-}
-
-// Description implements sql.FunctionExpression
-func (d *DateSub) Description() string {
-	return "subtracts the interval from the given date."
-}
-
-// Children implements the sql.Expression interface.
-func (d *DateSub) Children() []sql.Expression {
-	return []sql.Expression{d.Date, d.Interval}
-}
-
-// Resolved implements the sql.Expression interface.
-func (d *DateSub) Resolved() bool {
-	return d.Date.Resolved() && d.Interval.Resolved()
-}
-
-// IsNullable implements the sql.Expression interface.
-func (d *DateSub) IsNullable() bool {
-	return true
-}
-
-// Type implements the sql.Expression interface.
-func (d *DateSub) Type() sql.Type {
-	sqlType := dateOffsetType(d.Date, d.Interval)
-	return sqlType
-}
-
-// CollationCoercibility implements the interface sql.CollationCoercible.
-func (*DateSub) CollationCoercibility(ctx *sql.Context) (collation sql.CollationID, coercibility byte) {
-	return ctx.GetCollation(), 4
-}
-
-// WithChildren implements the Expression interface.
-func (d *DateSub) WithChildren(children ...sql.Expression) (sql.Expression, error) {
-	return NewDateSub(children...)
-}
-
-// Eval implements the sql.Expression interface.
-func (d *DateSub) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
-	date, err := d.Date.Eval(ctx, row)
-	if err != nil {
-		return nil, err
-	}
-
-	if date == nil {
-		return nil, nil
-	}
-
-	delta, err := d.Interval.EvalDelta(ctx, row)
-	if err != nil {
-		return nil, err
-	}
-
-	if delta == nil {
-		return nil, nil
-	}
-
-	var dateVal interface{}
-	dateVal, _, err = types.DatetimeMaxPrecision.Convert(ctx, date)
-	if err != nil {
-		ctx.Warn(1292, err.Error())
-		return nil, nil
-	}
-
-	// return appropriate type
-	res := types.ValidateTime(delta.Sub(dateVal.(time.Time)))
-	if res == nil {
-		return nil, nil
-	}
-
-	resType := d.Type()
-	if types.IsText(resType) {
-		// If the input is a properly formatted date/datetime string, the output should also be a string
-		if dateStr, isStr := date.(string); isStr {
-			if res.(time.Time).Nanosecond() > 0 {
-				return res.(time.Time).Format(sql.DatetimeLayoutNoTrim), nil
-			}
-			if isHmsInterval(d.Interval) {
-				return res.(time.Time).Format(sql.TimestampDatetimeLayout), nil
-			}
-			for _, layout := range types.DateOnlyLayouts {
-				if _, pErr := time.Parse(layout, dateStr); pErr != nil {
-					continue
-				}
-				return res.(time.Time).Format(sql.DateLayout), nil
-			}
-		}
-	}
-
-	ret, _, err := resType.Convert(ctx, res)
-	if err != nil {
-		return nil, err
-	}
-	return ret, nil
-}
-
-func (d *DateSub) String() string {
-	return fmt.Sprintf("%s(%s,%s)", d.FunctionName(), d.Date, d.Interval)
-}
-
 // DatetimeConversion is a shorthand function for CONVERT(expr, DATETIME)
 type DatetimeConversion struct {
 	Date sql.Expression
